+++ R/prob-roc_aunu.R
@@ -5,18 +5,6 @@
 #' curve of each class against the rest, using the uniform class distribution.
 #' This is equivalent to `roc_auc(estimator = "macro")`.
 #'
-#' Like the other ROC AUC metrics, `roc_aunu()` defaults to allowing
-#' `pROC::roc()` control the direction of the computation, but allows you to
-#' control this by passing `options = list(direction = "<")` or any other
-#' allowed direction value. pROC advises setting the `direction` when doing
-#' resampling so that the AUC values are not biased upwards.
-#'
-#' Generally, an ROC AUC value is between `0.5` and `1`, with `1` being a
-#' perfect prediction model. If your value is between `0` and `0.5`, then
-#' this implies that you have meaningful information in your model, but it
-#' is being applied incorrectly because doing the opposite of what the model
-#' predicts would result in an AUC `>0.5`.
-#'
 #' @family class probability metrics
 #' @templateVar metric_fn roc_aunu
 #' @template return
@@ -27,7 +15,7 @@
 #' uniform class distribution and is equivalent to
 #' `roc_auc(estimator = "macro")`.
 #'
-#' @inheritParams pr_auc
+#' @inheritParams roc_auc
 #'
 #' @param ... A set of unquoted column names or one or more `dplyr` selector
 #' functions to choose which variables contain the class probabilities. There
@@ -36,10 +24,6 @@
 #' @param estimate A matrix with as many
 #' columns as factor levels of `truth`. _It is assumed that these are in the
 #' same order as the levels of `truth`._
-#'
-#' @param options A `list` of named options to pass to [pROC::roc()]
-#' such as `direction` or `smooth`. These options should not include `response`,
-#' `predictor`, `levels`, or `quiet`.
 #'
 #' @references
 #'
@@ -109,15 +93,18 @@
 roc_aunu <- function(data, ...) {
   UseMethod("roc_aunu")
 }
-
-class(roc_aunu) <- c("prob_metric", "function")
-attr(roc_aunu, "direction") <- "maximize"
+roc_aunu <- new_prob_metric(
+  roc_aunu,
+  direction = "maximize"
+)
 
 #' @export
 #' @rdname roc_aunu
-roc_aunu.data.frame  <- function(data, truth, ..., options = list(),
+roc_aunu.data.frame  <- function(data,
+                                 truth,
+                                 ...,
+                                 options = list(),
                                  na_rm = TRUE) {
-
   estimate <- dots_to_estimate(data, !!! enquos(...))
 
   metric_summarizer(
@@ -128,23 +115,33 @@ roc_aunu.data.frame  <- function(data, truth, ..., options = list(),
     estimate = !!estimate,
     estimator = NULL,
     na_rm = na_rm,
+    event_level = NULL,
     ... = ...,
     metric_fn_options = list(options = options)
   )
-
 }
 
 #' @rdname roc_aunu
 #' @export
 #' @importFrom rlang call2
 #' @importFrom pROC roc auc
-roc_aunu_vec <- function(truth, estimate, options = list(),
-                         na_rm = TRUE, ...) {
-
+roc_aunu_vec <- function(truth,
+                         estimate,
+                         options = list(),
+                         na_rm = TRUE,
+                         ...) {
   estimator <- "macro"
 
+  # `event_level` doesn't really matter, but we set it anyways
   roc_aunu_impl <- function(truth, estimate) {
-    roc_auc_vec(truth, estimate, options, estimator)
+    roc_auc_vec(
+      truth = truth,
+      estimate = estimate,
+      options = options,
+      estimator = estimator,
+      na_rm = FALSE,
+      event_level = "first"
+    )
   }
 
   metric_vec_template(